// src/model/pmtkTypes.ts
import type { EventCategory } from "./eventcategory.ts";

export interface tEvents {
  ID: number;
  SeriesID: number;
  EventName: string;
  EventDate: Date;
  EventLocation: string;
  EventDistance: number;
  EventDuration: number;
  EventLaps: number;
}

export interface tEventsRiders {
  ID: number;
  EventID: number;
  EventRaceNo: number;
  TagNo: number;
  CategoryCode: string;
  RiderID: number;
  TimingAdjustment: number;
  TeamID: number;
  SpecialStatus: string;
}

export interface tRiders {
  ID: number;
  CANo: number;
  MTBANo: number;
  IMBANo: number;
  BMXANo: number;
  FirstName: string;
  Surname: string;
  DisplayName: string;
  AgeGroupCode: string;
  Address: string;
  Suburb: string;
  State: string;
  PostCode: string;
  Country: string;
  PhoneNo: string;
  Mobile: string;
  email: string;
  ToReceiveMailOuts: string;
  Gender: string;
  DOB: Date;
  NextOfKin: string;
  EmergencyName: string;
  EmergencyPhone: string;
}

export interface tEventRidersResultsCommon {
  ID: number;
  EventID: number;
  EventRaceNo: number;
  TagNo: number;
  CrossLineAtDT: Date;
  CrossLineAt: string;
  TotalRideTime: string;
  SplitTime: string;
  PosNo: number;
}

export interface tEventRidersResults extends tEventRidersResultsCommon {
  IsIgnoring: false;
}

export interface tEventRidersResults_Ignoring extends tEventRidersResultsCommon {
  ReasonTest: string;
  IsIgnoring: true;
}

export interface PMTKLookupEventCategory extends EventCategory {
  categoryCode: string;
}

export interface tChipTimes {
  ID: number;
  Location: string;
  ChipCode: string;
  ChipTime: Date;
  ReaderTime: Date;
  Milliseconds: number;
  ReaderNo: number;
  AntennaNo: number;
}

export type PMTKRiderResultTypes = tEventRidersResults | tEventRidersResults_Ignoring;